requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-index
#
certifi==2020.4.5.1       # via requests
cffi==1.14.0              # via cryptography
chardet==3.0.4            # via requests
cryptography==2.9.2       # via pyopenssl, requests
enum34==1.1.10            # via cryptography
idna==2.9                 # via requests
ipaddress==1.0.23         # via cryptography
kafka-python==2.0.1
psycopg2-binary==2.8.5
pycparser==2.20           # via cffi
pyopenssl==19.1.0         # via requests
python-snappy==0.5.4
requests[security]==2.23.0
schematics==2.1.0
six==1.15.0               # via cryptography, pyopenssl
urllib3==1.25.9           # via requests
